Updating testing scripts.
streamlining passwords for current scripts.
Adding new script to form questions that make it easy to visually inspect for inconsistencies.
Next step should be adding likes and also, for the sake of testing potential bugs, deleting questions / quizzes / justifications / distractors then re-inserting them so as to make sure we do not confuse at some point in the code the IDs of these entities with their index in some array.