Uninspired Ramblings Added.
Added some comments at the end of step2.sh to illustrate what the values for maxlikes should be with the current test scripts.
We need to increase the number of accounts in the test scripts and the number of likes given by students to one another so that we may have something a bit more reaslistic.
For instance, some alternatives should have more than 2 other justifications available to show to any one student.
This will, in turn, allow the likes received to go over 2 for any justification.
Remember that our code tries to find at max 3 of them to show, or less is less are available.
</duplicatedramblings>